requirements.txt
airtable-python-wrapper==0.11.3.post1
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pylint==2.3.1
requests==2.22.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.25.3
wrapt==1.11.1
